Renames orp_search to search. Renames templates to reflect Django and React based searches
