+++ fbr/search/legislation.py
@@ -6,12 +6,12 @@ from typing import Optional
 
 import requests  # type: ignore
 
-from orp_search.config import SearchDocumentConfig
-from orp_search.construction_legislation import (  # noqa: E501
+from search.config import SearchDocumentConfig
+from search.construction_legislation import (  # noqa: E501
     construction_legislation_dataframe,
 )
-from orp_search.utils.date import convert_date_string_to_obj
-from orp_search.utils.documents import (  # noqa: E501
+from search.utils.date import convert_date_string_to_obj
+from search.utils.documents import (  # noqa: E501
     generate_short_uuid,
     insert_or_update_document,
 )
