+++ react_front_end/src/index.js
@@ -2,7 +2,7 @@ import React from "react"
 import ReactDOM from "react-dom/client"
 import App from "./App"
 
-const root = ReactDOM.createRoot(document.getElementById("orp-react-app"))
+const root = ReactDOM.createRoot(document.getElementById("find-business-regulations-react-app"))
 root.render(
   <React.StrictMode>
     <App />
